Server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appJar==0.94.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.1
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
ecdsa==0.18.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.14.0
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gotrue==2.4.2
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.2
ipython==8.21.0
ipython_genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-console==6.4.0
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==23.2
paho-mqtt==2.0.0
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.3
PeakUtils==1.3.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.2.0
ply==3.11
postgrest==0.16.4
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
realtime==1.0.4
referencing==0.33.0
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
stmpy==0.7.5
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
sympy==1.9
systemd-python==234
terminado==0.18.0
testpath==0.5.0
thread6==0.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
vigenerecipher==0.3
virtualenv==20.26.1
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.10
xdg==5
xkit==0.0.0
zipp==1.0.0
